Move DC dispatcher requests and responses to ui module
