+++ example/src/main/kotlin/com/processout/example/ui/screen/checkout/DefaultDynamicCheckoutDelegate.kt
@@ -2,7 +2,7 @@ package com.processout.example.ui.screen.checkout
 
 import com.processout.example.shared.Constants
 import com.processout.sdk.api.model.request.POCreateInvoiceRequest
-import com.processout.sdk.api.model.request.PODynamicCheckoutInvoiceInvalidationReason
+import com.processout.sdk.ui.checkout.delegate.PODynamicCheckoutInvoiceInvalidationReason
 import com.processout.sdk.api.model.request.POInvoiceRequest
 import com.processout.sdk.api.model.response.POInvoice
 import com.processout.sdk.api.service.POInvoicesService
