+++ sdk/src/main/kotlin/com/processout/sdk/api/model/response/PONativeAlternativePaymentMethodDefaultValuesResponse.kt
@@ -1,7 +1,6 @@
 package com.processout.sdk.api.model.response
 
 import com.processout.sdk.api.dispatcher.POEventDispatcher
-import com.processout.sdk.api.model.request.PODynamicCheckoutAlternativePaymentDefaultValuesRequest
 import com.processout.sdk.api.model.request.PONativeAlternativePaymentMethodDefaultValuesRequest
 import com.processout.sdk.core.annotation.ProcessOutInternalApi
 import java.util.UUID
@@ -13,7 +12,7 @@ import java.util.UUID
  * @param[uuid] Unique identifier of response that must be equal to UUID of request.
  * @param[defaultValues] Map where key is [PONativeAlternativePaymentMethodParameter.key] and value is a default value for this parameter.
  */
-data class PONativeAlternativePaymentMethodDefaultValuesResponse internal constructor(
+data class PONativeAlternativePaymentMethodDefaultValuesResponse @ProcessOutInternalApi constructor(
     override val uuid: UUID,
     val defaultValues: Map<String, String>
 ) : POEventDispatcher.Response
@@ -24,11 +23,5 @@ data class PONativeAlternativePaymentMethodDefaultValuesResponse internal constr
  * @param[defaultValues] Map where key is [PONativeAlternativePaymentMethodParameter.key] and value is a default value for this parameter.
  */
 fun PONativeAlternativePaymentMethodDefaultValuesRequest.toResponse(
-    defaultValues: Map<String, String>
-) = PONativeAlternativePaymentMethodDefaultValuesResponse(uuid, defaultValues)
-
-/** @suppress */
-@ProcessOutInternalApi
-fun PODynamicCheckoutAlternativePaymentDefaultValuesRequest.toResponse(
     defaultValues: Map<String, String>
 ) = PONativeAlternativePaymentMethodDefaultValuesResponse(uuid, defaultValues)
