+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/delegate/PODynamicCheckoutAlternativePaymentDefaultValuesRequest.kt
@@ -1,7 +1,8 @@
-package com.processout.sdk.api.model.request
+package com.processout.sdk.ui.checkout.delegate
 
 import com.processout.sdk.api.dispatcher.POEventDispatcher
 import com.processout.sdk.api.model.response.PODynamicCheckoutPaymentMethod
+import com.processout.sdk.api.model.response.PONativeAlternativePaymentMethodDefaultValuesResponse
 import com.processout.sdk.api.model.response.PONativeAlternativePaymentMethodParameter
 import com.processout.sdk.core.annotation.ProcessOutInternalApi
 import java.util.UUID
@@ -13,3 +14,9 @@ data class PODynamicCheckoutAlternativePaymentDefaultValuesRequest(
     val paymentMethod: PODynamicCheckoutPaymentMethod.AlternativePayment,
     val parameters: List<PONativeAlternativePaymentMethodParameter>
 ) : POEventDispatcher.Request
+
+/** @suppress */
+@ProcessOutInternalApi
+fun PODynamicCheckoutAlternativePaymentDefaultValuesRequest.toResponse(
+    defaultValues: Map<String, String>
+) = PONativeAlternativePaymentMethodDefaultValuesResponse(uuid, defaultValues)
