+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/delegate/PODynamicCheckoutInvoiceAuthorizationResponse.kt
@@ -1,7 +1,6 @@
-package com.processout.sdk.api.model.response
+package com.processout.sdk.ui.checkout.delegate
 
 import com.processout.sdk.api.dispatcher.POEventDispatcher
-import com.processout.sdk.api.model.request.PODynamicCheckoutInvoiceAuthorizationRequest
 import com.processout.sdk.api.model.request.POInvoiceAuthorizationRequest
 import com.processout.sdk.core.annotation.ProcessOutInternalApi
 import java.util.UUID
