+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/delegate/PODynamicCheckoutInvoiceRequest.kt
@@ -1,4 +1,4 @@
-package com.processout.sdk.api.model.request
+package com.processout.sdk.ui.checkout.delegate
 
 import com.processout.sdk.api.dispatcher.POEventDispatcher
 import com.processout.sdk.api.model.response.POInvoice
