+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/delegate/PODynamicCheckoutInvoiceResponse.kt
@@ -1,8 +1,6 @@
-package com.processout.sdk.api.model.response
+package com.processout.sdk.ui.checkout.delegate
 
 import com.processout.sdk.api.dispatcher.POEventDispatcher
-import com.processout.sdk.api.model.request.PODynamicCheckoutInvoiceInvalidationReason
-import com.processout.sdk.api.model.request.PODynamicCheckoutInvoiceRequest
 import com.processout.sdk.api.model.request.POInvoiceRequest
 import com.processout.sdk.core.annotation.ProcessOutInternalApi
 import java.util.UUID
